move tracelistener to its own file.
